pyproject.toml
[project]
name = "MomokoBot"
version = "0.1.0"
description = "MomokoBot"
readme = "README.md"
requires-python = ">=3.9, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = [
    "basic_plugins",
    "plugins",
    "other_plugins"
]

[tool.poetry]
name = "MomokoBot"
version = "0.1.0"
description = "MomokoBot"
authors = ["月ヶ瀬"]
license = "GPL-3.0"
readme = "README.md"
homepage = "https://github.com/tkgs0/Momoko"
repository = "https://github.com/tkgs0/Momoko"
keywords = ["nonebot", "momoko", "qqbot"]

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
nb-cli = ">=1.4.1"
nonebot2 = {extras = ["fastapi"], version = "^2.2.1"}
nonebot-adapter-onebot = "^2.4.1"
nonebot-plugin-apscheduler = ">=0.2.0"
nonebot-plugin-imageutils = ">=0.1.13"
nonebot-plugin-guild-patch = ">=0.2.0"
nonebot-plugin-send-anything-anywhere = [{ version = ">=0.3.2" }]
aiofiles = ">=22.1.0"
aiohttp = {extras = ["speedups"], version = ">=3.8.3"}
cachetools = ">=5.0.0"
deep-translator = ">=1.8.0"
diro-py = ">=0.1.0"
diskcache = ">=5.4.0"
emoji = ">=2.0.0"
feedparser = ">=6.0.3"
imagehash = ">=4.3.0"
langid = ">=1.1.6"
magneturi = ">=1.3"
picimagesearch = ">=3.9.2"
psutil = ">=5.9.0"
pydub = ">=0.25.1"
python-qbittorrent = ">=0.4.1"
shelved-cache = ">=0.3.1"
tenacity = ">=8.0.0"
tinydb = ">=4.6.0"
ujson = ">=5.2.0"
arrow = ">=1.2.3"
dbutils = ">=3.0.2"
pikpakapi = "^0.1.6"

[tool.poetry.dev-dependencies]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
implicit_reexport = true
pretty = true
show_error_codes = true
strict = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

